requirements.txt
ccxt
PySide6
python-dotenv

# Официальные SDK (стабильные)
pybit
binance-connector
python-okx